requirements.txt
# Django version
django>=1.7

pyelasticsearch

# API
djangorestframework
#django-oauth-toolkit
djangorestframework-jwt

# Testing
factory_boy
coverage
selenium
django-nose
coveralls
fake-factory
django-debug-toolbar
